refactor(test): extract shared enrichment claimRow helper (BS#892)
The claim CAS-UPDATE mirror of apps/enrichment-worker/claim.ts was byte-identical in three integration specs (claim, sweep, backpressure), so a hand-edit to claim.ts had to be chased through three copies. Move it to tests/utils/enrichment-claim.js as the single canonical mirror; all three specs import it. Canonicalize on the {claimed, id} return shape and adapt the sweep spec's boolean call sites.

The per-spec insert helpers stay local by design: genuinely different insert strategies (single-row VALUES with a readable suffix vs. bulk generate_series) and distinct play_order sentinels chosen so siblings can't collide under the shared --runInBand schema — folding them would be speculative generality.
